refactor: Delete UserUtils - dto, Entity 변환과 객체 생성은 dto class 이동
